Fix IDE nag
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1807751 13f79535-47bb-0310-9956-ffa450edef68

+++ test/org/apache/tomcat/util/net/openssl/TestOpenSSLConf.java
@@ -69,7 +69,7 @@ public class TestOpenSSLConf extends TomcatBaseTest {
         String[] ciphers = sslHostConfig.getEnabledCiphers();
         assertEquals("Checking enabled cipher count", 1, ciphers.length);
         assertEquals("Checking enabled cipher", CIPHER, ciphers[0]);
-        ciphers = SSLContext.getCiphers(sslHostConfig.getOpenSslContext());
+        ciphers = SSLContext.getCiphers(sslHostConfig.getOpenSslContext().longValue());
         assertEquals("Checking context cipher count", 1, ciphers.length);
         assertEquals("Checking context cipher", CIPHER, ciphers[0]);
     }
